Fixes how examine starts up to make it easier for custom index developers
